+++ src/edu/umass/cs/gnsserver/gnsapp/clientCommandProcessor/commands/select/Select.java
@@ -38,6 +38,7 @@ import org.json.JSONObject;
  *
  * @author westy
  */
+@Deprecated
 public class Select extends AbstractCommand {
 
   /**
