+++ src/edu/umass/cs/gnsserver/gnsapp/clientCommandProcessor/commands/select/SelectWithin.java
@@ -38,6 +38,7 @@ import org.json.JSONObject;
  *
  * @author westy
  */
+@Deprecated
 public class SelectWithin extends AbstractCommand {
 
   /**
